LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssUnparsedValue-length-expected.txt


FAIL Length of CSSUnparsedValue with no fragments is zero assert_equals: length expected (number) 0 but got (undefined) undefined
FAIL Length of CSSUnparsedValue with multiple fragments is the number of fragments Can't find variable: CSSVariableReferenceValue
FAIL Length of CSSUnparsedValue updates when fragments are appended assert_equals: initial length expected (number) 1 but got (undefined) undefined
FAIL Length of CSSUnparsedValue does not change when fragments are modified assert_equals: initial length expected (number) 1 but got (undefined) undefined
